feat(#219): use storage in DefaultDecompiler
